[#20] Divide result failure and success page for convenient routing
