+++ src/index.js
@@ -13,45 +13,46 @@ import Company from './pages/Company'
 import { Introduction, Activity, Training, Benefit, Office } from './pages/Company/'
 
 import Game from './pages/Game'
-import { Questions, Result, Ranking } from './pages/Game/'
+import { Questions, ResultSuccess, ResultFailure, Ranking } from './pages/Game/'
 
 import NotFound from './pages/NotFound'
 
 import './index.css'
 
 ReactDOM.render((
-    <Router history={hashHistory}>
+   <Router history={hashHistory}>
       <Route path="/">
-        <IndexRoute component={Main}/>
-        <Route path="recruit">
-          <IndexRoute component={Recruit}/>
-          <Route path="schedule" component={Schedule}/>
-          <Route path="position" component={Position}/>
-          <Route path="announce" component={Announce}/>
-          <Route path="faq" component={FAQ}/>
-        </Route>
-        <Route path="campus">
-          <IndexRoute component={Campus}/>
-          <Route path="coding" component={Coding}/>
-          <Route path="openday" component={OpenDay}/>
-          <Route path="girl" component={Girl}/>
-        </Route>
-        <Route path="company">
-          <IndexRoute component={Company}/>
-          <Route path="introduction" component={Introduction}/>
-          <Route path="activity" component={Activity}/>
-          <Route path="training" component={Training}/>
-          <Route path="benefit" component={Benefit}/>
-          <Route path="office" component={Office}/>
-        </Route>
-        <Route path="game">
-          <IndexRoute component={Game}/>
-           <Route path="questions" component={Questions} />
-          <Route path="result" component={Result}/>
-           <Route path="ranking" component={Ranking} />
-        </Route>
-        <Route path="*" component={NotFound}/>
+         <IndexRoute component={Main}/>
+         <Route path="recruit">
+            <IndexRoute component={Recruit}/>
+            <Route path="schedule" component={Schedule}/>
+            <Route path="position" component={Position}/>
+            <Route path="announce" component={Announce}/>
+            <Route path="faq" component={FAQ}/>
+         </Route>
+         <Route path="campus">
+            <IndexRoute component={Campus}/>
+            <Route path="coding" component={Coding}/>
+            <Route path="openday" component={OpenDay}/>
+            <Route path="girl" component={Girl}/>
+         </Route>
+         <Route path="company">
+            <IndexRoute component={Company}/>
+            <Route path="introduction" component={Introduction}/>
+            <Route path="activity" component={Activity}/>
+            <Route path="training" component={Training}/>
+            <Route path="benefit" component={Benefit}/>
+            <Route path="office" component={Office}/>
+         </Route>
+         <Route path="game">
+            <IndexRoute component={Game}/>
+            <Route path="questions" component={Questions}/>
+            <Route path="result/failure" component={ResultFailure} />
+            <Route path="result/success" component={ResultSuccess} />
+            <Route path="ranking" component={Ranking}/>
+         </Route>
+         <Route path="*" component={NotFound}/>
       </Route>
-    </Router>
-  ), document.getElementById('root')
+   </Router>
+), document.getElementById('root')
 )
