+++ src/pages/Game.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { Link } from 'react-router'
 
 import './Game/Questions.less'
